fix: renamed aggregate components
