+++ Source/Extensions/Blazorise.DataGrid/DataGrid.razor
@@ -104,7 +104,7 @@
         @if ( HasAggregates )
         {
             <TableFooter>
-                <_DataGridAggregateRow TItem="TItem" AggregateColumns="@AggregateColumns" Columns="@Columns" Class="@GroupRowClass" Style="@GroupRowStyle" />
+                <_DataGridAggregateRow TItem="TItem" Aggregates="@Aggregates" Columns="@Columns" Class="@GroupRowClass" Style="@GroupRowStyle" />
             </TableFooter>
         }
     </Table>
